@@ -1,71 +1,36 @@
 {
-  "name": "lunch",
+  "name": "workremotely.asia",
   "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "start": "env BABEL_ENV=server babel-watch server/server.js",
-    "test": "jest test --config test/jestSetup.js",
-    "test-tdd": "jest test --config test/jestSetup.js --watch",
-    "jest-update": "jest --updateSnapshot --config test/jestSetup.js",
-    "webpack-dev": "webpack-dev-server --host 0.0.0.0 --progress --hot --config config/webpack.dev.js",
-    "dev": "npm run start & npm run webpack-dev"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/daisyxu0924/where-for-lunch.git"
-  },
-  "keywords": [],
-  "author": "daisyxu0924",
-  "license": "ISC",
+  "description": "Personal Project",
+  "private": true,
   "dependencies": {
-    "axios": "^0.18.1",
-    "babel-polyfill": "^6.26.0",
-    "classnames": "^2.2.5",
-    "css-loader": "^0.28.11",
-    "postcss": "^6.0.22",
-    "postcss-advanced-variables": "^2.3.3",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.1.5",
-    "postcss-nested": "^3.0.0",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-hot-loader": "^4.0.0",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-router-redux": "^4.0.8",
-    "redux": "^4.0.0",
-    "redux-actions": "^2.4.0",
-    "redux-saga": "^0.16.0"
+    "@atlaskit/css-reset": "^5.0.2",
+    "axios": "^0.19.0",
+    "react": "^16.8.6",
+    "react-beautiful-dnd": "^11.0.4",
+    "react-dom": "^16.8.6",
+    "react-scripts": "3.0.1",
+    "styled-components": "^4.3.2"
   },
-  "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-watch": "^2.0.7",
-    "dotenv": "^5.0.1",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-mocha": "^5.0.0",
-    "eslint-plugin-react": "^7.8.2",
-    "eslint-plugin-standard": "^3.1.0",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.1.0",
-    "react-test-renderer": "^16.4.0",
-    "redux-devtools": "^3.4.1",
-    "style-loader": "^0.21.0",
-    "webpack": "^4.10.2",
-    "webpack-cli": "^2.1.4",
-    "webpack-dev-server": "^3.1.4"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
