@@ -1,22 +1,23 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 
-export default class remoteOK extends React.Component{
-  componentDidMount(){
-    axios.get('https://remoteok.io/api')
-    .then(resp=> {
-      const feed= resp.data;
-      this.setState({feed});
-    })
-  }
-  render(){
-    return (
+const remoteOK = () => {
+  const [feed, setFeed] = useState([]);
+  useEffect(() => {
+    const fetchData = await () => {
+      const response = await axios.get('https://remoteok.io/api');
+      setFeed(response.data);
+    };
+    fetchData();
+  }, []);
+  
+  return (
       <h1>Feed</h1>
 
       <ul>
         <li>
           <p class="legal">{feed.legal}</p>
-          <a class="position" routerLink="/details/{{feed.id}}">{feed.position }</a>
+          <a class="position" href="#">{feed.position }</a>
           
           <ul>
           <li class="company">{ feed.company }</li>
@@ -27,6 +28,6 @@ export default class remoteOK extends React.Component{
         </li>
       </ul>
     )
-  }
 }
-export default remoteOK ;
+
+export default remoteOK;
